// src/redux/licenses/action-types.ts
export enum ActionTypes {
    ADD_ANSWER = "ADD_ANSWER",
    ADD_FIELDS = "ADD_FIELDS",
    ADD_IMAGE = "ADD_IMAGE",
    ADD_QUESTION = "ADD_QUESTION",
    ADD_SECTION = "ADD_SECTION",
    COMPLETE_REG_PROGRESS = "COMPLETE_REG_PROGRESS",
    CONFIRM_PAYMENT = "CONFIRM_PAYMENT",
    CREATE_LICENCE = "CREATE_LICENCE",
    DELETE_ANSWER = "DELETE_ANSWER",
    DELETE_IMAGE = "DELETE_IMAGE",
    DELETE_QUESTION = "DELETE_QUESTION",
    DELETE_SECTION = "DELETE_SECTION",
    DELETE_SLIDE_PROP = "DELETE_SLIDE_PROP",
    EDIT_ANSWER = "EDIT_ANSWER",
    EDIT_QUESTION = "EDIT_QUESTION",
    EDIT_SECTION = "EDIT_SECTION",
    GET_LICENCES = "GET_LICENCES",
    GET_SEARCH_LICENSE = "GET_SEARCH_LICENSE",
    GET_USER = "GET_USER",
    OVERVIEW_COURSE = "OVERVIEW_COURSE",
    RESET_ADMIN_LICENSE_STATE = "RESET_ADMIN_LICENSE_STATE",
    SET_COLLAPSE_STATUS = "SET_COLLAPSE_STATUS",
    SET_CURRENT_LICENSE_NAME = "SET_CURRENT_LICENSE_NAME",
    SET_CURRENT_SLIDE_POSITION = "SET_CURRENT_SLIDE_POSITION",
    SET_ERROR_STATUS = "SET_ERROR_STATUS",
    SET_GET_LICENSE_ERROR = "SET_GET_LICENSE_ERROR",
    START_GET_LICENSE = "START_GET_LICENSE",
    START_QUIZ = "START_QUIZ",
    START_SLIDE_SHOW = "START_SLIDE_SHOW",
    USER_REGISTER = "USER_REGISTER"
}